Few changes to keep devtools document() and R CMD check happy.
I need an export for my classes in the namespace. By default,
document() generates an exportClasses. Problem with exportClasses is
that my classes are not visible. When an export for a class is created,
related .rd files have to be renamed (class suffix has to be removed).
Otherwise check issues a warning…